Renames the `inc` folder into a `logue` folder allowing for global-style imports.
