+++ platform/minilogue-xd/logue/dsp/biquad.hpp
@@ -41,7 +41,7 @@
  *
  */
 
-#include "float_math.h"
+#include "../utils/float_math.h"
 
 /**
  * Common DSP Utilities
