+++ platform/prologue/logue/osc_api.h
@@ -44,9 +44,9 @@
 #ifndef __osc_api_h
 #define __osc_api_h
 
-#include "float_math.h"
-#include "int_math.h"
-#include "fixed_math.h"
+#include "utils/float_math.h"
+#include "utils/int_math.h"
+#include "utils/fixed_math.h"
 
 #ifdef __cplusplus
 extern "C" {
