+++ platform/prologue/logue/userdelfx.h
