+++ platform/prologue/logue/utils/buffer_ops.h
