+++ platform/prologue/logue/utils/fixed_math.h
