+++ platform/prologue/logue/utils/int_math.h
